requirements.txt
requests==2.25.0
flask==1.1.2
python-dotenv==0.15.0
tenacity==6.2.0
